+++ concordium-consensus/src/Concordium/Queries.hs
@@ -659,7 +659,7 @@ getNextAccountNonce accountAddress =
         )
         -- consensus v1
         ( do
-            (nanNonce, nanAllFinal) <- gets (SkovV1.getNextAccountNonce acctEq)
+            (nanNonce, nanAllFinal) <- SkovV1.getNextAccountNonce acctEq =<< get
             return NextAccountNonce{..}
         )
   where